google-auth-fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
authlib==1.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2